% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{build_etl_job_log_df}
\alias{build_etl_job_log_df}
\title{Builds formatted rcc_job_log data frame from source data}
\usage{
build_etl_job_log_df(job_duration, job_summary, level)
}
\arguments{
\item{job_duration, }{the job duration in seconds}

\item{job_summary, }{the summary of the job performed}

\item{level, }{the log level (DEBUG, ERROR, INFO)}
}
\value{
df_etl_log_job, the df matching the rcc_job_log table
}
\description{
Builds formatted rcc_job_log data frame from source data
}
\examples{
\dontrun{
 build_etl_job_log_df(
   job_duration,
   job_summary,
   level
 )
}
}
