% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{build_formatted_df_from_result}
\alias{build_formatted_df_from_result}
\title{Builds formatted data frame from source data}
\usage{
build_formatted_df_from_result(
  result,
  database_written,
  table_written,
  log_level,
  pk_col
)
}
\arguments{
\item{result, }{the df to log}

\item{database_written, }{the database the etl wrote to}

\item{table_written, }{the table the etl wrote to}

\item{log_level, }{the log level (DEBUG, ERROR, INFO)}

\item{pk_col, }{the dataframe col to use as the primary_key}
}
\value{
df_etl_log, the df matching the etl_log table
}
\description{
Builds formatted data frame from source data
}
\examples{
\dontrun{
 build_formatted_df_from_result(
   result,
   database_written,
   table_written,
   log_level,
   pk_col
 )
}
}
