% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to db}
\usage{
connect_to_db(drv, prefix = "", continue_on_error = FALSE)
}
\arguments{
\item{drv, }{an object that inherits from DBIDriver (e.g. RMariaDB::MariaDB()), or an existing DBIConnection object (in order to clone an existing connection).}

\item{prefix, }{the}

\item{continue_on_error}{if TRUE then continue execution on error, if FALSE then quit non interactive sessions on error}
}
\value{
An S4 object. Run ?dbConnect for more information
}
\description{
Connect to db
}
\examples{
\dontrun{
# connect to db using [prefix]_DB_* environment variables
con <- connect_to_db(drv = RMariaDB::MariaDB(), prefix = "RCC")

# connect to sqlite db
con <- connect_to_db(drv = RSQLite::SQLite())
}
}
