% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{connect_to_log_db}
\alias{connect_to_log_db}
\title{Connect to the log db}
\usage{
connect_to_log_db(drv, continue_on_error = FALSE)
}
\arguments{
\item{drv, }{an object that inherits from DBIDriver (e.g. RMariaDB::MariaDB()), or an existing DBIConnection object (in order to clone an existing connection).}

\item{continue_on_error}{if TRUE then continue execution on error, if FALSE then quit non interactive sessions on error}
}
\value{
An S4 object. Run ?dbConnect for more information
}
\description{
Connect to the log db
}
\examples{

\dontrun{
# connect to log db using LOG_DB_* environment variables
con <- connect_to_log_db()

# connect to sqlite log db
con <- connect_to_log_db(drv = RSQLite::SQLite())
}
}
