% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{convert_schema_to_sqlite}
\alias{convert_schema_to_sqlite}
\title{Converts a MySQL schema file to a sqlite schema.
Facilitates easier creation of in-memory (i.e. sqlite) tables.}
\usage{
convert_schema_to_sqlite(schema_file_path)
}
\arguments{
\item{schema_file_path, }{the path of the schema file to convert}
}
\value{
the translated schema as a character string
}
\description{
Converts a MySQL schema file to a sqlite schema.
Facilitates easier creation of in-memory (i.e. sqlite) tables.
}
\examples{
\dontrun{
mem_conn <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")
translated_schema <- convert_schema_to_sqlite("~/documents/my_cool_schema.sql")
DBI::dbSendQuery(mem_conn, schema)
}
}
