% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{create_test_table}
\alias{create_test_table}
\title{Create a test table from package schema and data files}
\usage{
create_test_table(conn, table_name, data_file = NA_character_, empty = F)
}
\arguments{
\item{conn}{A DBI Connection object}

\item{table_name}{The name of a table used in testing}

\item{data_file}{The name of a file with alternative contents for `table_name`}

\item{empty}{A boolean to request that the table be created empty}
}
\value{
NA
}
\description{
Creates tables from files in inst/testdata that match the patterns <table_name>_schema.sql and <table_name>.csv
}
\examples{
\dontrun{
conn <- dbConnect(RSQLite::SQLite(), dbname = ":memory:")
create_test_table(conn, "redcap_user_information")

}
}
