% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{create_test_tables}
\alias{create_test_tables}
\title{A wrapper around \code{\link{create_test_table}} to create all tables, or a specified subset of them}
\usage{
create_test_tables(conn, table_names = c())
}
\arguments{
\item{conn}{A DBI Connection object}

\item{table_names}{A character list of the names of all tables you wish
to create, if nothing is provided, the result of
\code{\link{get_test_table_names}} will be used to create all test tables}
}
\value{
NA
}
\description{
A wrapper around \code{\link{create_test_table}} to create all tables, or a specified subset of them
}
\examples{
\dontrun{
conn <- dbConnect(RSQLite::SQLite(), dbname = ":memory:")
create_test_tables(conn) # create all test tables

}
}
