% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_diff_test_user_data}
\alias{dataset_diff_test_user_data}
\title{dataset_diff_test_user_data}
\format{
A list of 5 variables:
\describe{
  \item{\code{source}}{tibble of the source data}
  \item{\code{source_pk}}{character primary key of the source data}
  \item{\code{target}}{tibble of the target data}
  \item{\code{target_pk}}{character primary key of the target data}
  \item{\code{result}}{list of dataset_diff output}
}
}
\usage{
dataset_diff_test_user_data
}
\description{
A list of test inputs and outputs for dataset_diff
}
\details{
DETAILS
}
\keyword{datasets}
