% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_project.R
\name{delete_project}
\alias{delete_project}
\title{Delete a Project from REDCap}
\usage{
delete_project(project_id, conn)
}
\arguments{
\item{project_id}{A project ID or vector of project IDs to be deleted.}

\item{conn}{A DBI connection object to the database that holds the `redcap_projects`
and `redcap_log_event*` tables.}
}
\value{
A list containing:
\itemize{
  \item n: the number of projects deleted
  \item number_rows_logged: the number of rows logged for the deletion event
  \item project_ids_deleted: a vector of project IDs that were deleted
  \item data: a data frame with each input project_id and its status after trying to delete it
}
}
\description{
Deletes specified projects from the REDCap system by setting the `date_deleted` field.
It will also log the event in the appropriate `log_event_table` for each project.
}
\examples{
\dontrun{
conn <- DBI::dbConnect(...)
delete_project(c(1, 2, 3), conn)
}
}
