% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{enable_randomization_on_a_preconfigured_project_in_production}
\alias{enable_randomization_on_a_preconfigured_project_in_production}
\title{enable_randomization_on_a_preconfigured_project_in_production}
\usage{
enable_randomization_on_a_preconfigured_project_in_production(
  target_conn,
  target_project_id
)
}
\arguments{
\item{target_conn}{- a DBI connection object pointing at the REDCap
database that houses the target project.}

\item{target_project_id}{- The project ID of the REDCap project that
will receive the mirrored randomization data.}
}
\value{
A logical indicating success or failure of the operation
}
\description{
Turn on randomization in the target project but only if it has already
   been moved to production and randomization has been configured.
}
\examples{
\dontrun{
enable_randomization_on_a_preconfigured_project_in_production(
  target_conn = target_conn,
  target_project_id = target_project_id
)
}
}
