% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_database_comparison.R
\name{evaluate_checksums}
\alias{evaluate_checksums}
\title{evaluate_checksums}
\usage{
evaluate_checksums(source_checksums, target_checksums)
}
\arguments{
\item{source_checksums}{The output of `get_table_checksum()` for the
source MySQL database of a database copy operation}

\item{target_checksums}{The output of `get_table_checksum()` for the
target MySQL database of a database copy operation}
}
\value{
A dataframe with these columns:
\describe{
  \item{table}{MySQL table name}
  \item{checksum_source}{checksum of source table returned by "CHECKSUM TABLE <table_name>"}
  \item{checksum_target}{checksum of target table returned by "CHECKSUM TABLE <table_name>"}
  \item{elapsed_time_source}{elapsed time required for the query and fetch of the checksum of the source table}
  \item{elapsed_time_target}{elapsed time required for the query and fetch of the checksum of the target table}
  \item{matches}{a boolean indicating if the source and target checksums match}
}
}
\description{
Transform the output of `get_table_checksum()` from a source and target
copy of a MySQL database copy operation and compare the checksums.
}
\examples{
\dontrun{
evaluate_checksums(source_checksums, target_checksums)
}
}
