% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_rights.R
\name{expire_user_project_rights}
\alias{expire_user_project_rights}
\title{Expire user rights to REDCap projects}
\usage{
expire_user_project_rights(
  conn,
  project_ids,
  usernames = NULL,
  all_users_except = NULL,
  expiration_date = as.Date(NA)
)
}
\arguments{
\item{conn, }{DBI Connection object to a REDCap database}

\item{project_ids, }{a vector of project IDs whose users need expiration}

\item{usernames, }{a vector of usernames to expire across the vector of project IDs}

\item{all_users_except, }{a vector of usernames who will be excluded from expiration}

\item{expiration_date, }{the expiration date to be applied to the users. Defaults to today.}
}
\value{
a list of the update count and data written
\itemize{
  \item updates - the number of records revised
  \item data - the dataframe of changes applied to redcap_user_rights
}
}
\description{
Expire user rights on one or more REDCap projects based
on list of users to expire or users to exclude from expiration
}
\examples{
conn <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")

DBI::dbCreateTable(
  conn,
  name = "redcap_user_rights",
  fields = user_rights_test_data$redcap_user_rights
)
DBI::dbAppendTable(
  conn,
  name = "redcap_user_rights",
  value = user_rights_test_data$redcap_user_rights
)

usernames <- c("bob", "dan")

expire_user_project_rights(
  conn = conn,
  project_ids = c(34),
  usernames = usernames
)
}
