% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{export_allocation_tables_from_project}
\alias{export_allocation_tables_from_project}
\title{export_allocation_tables_from_project}
\usage{
export_allocation_tables_from_project(conn, project_id_to_export)
}
\arguments{
\item{conn}{- a DBI connection object pointing at a REDCap
database that houses the project on interest}

\item{project_id_to_export}{- The project ID of a REDCap project that
contains randomization to be exported.}
}
\value{
a dataframe in the shape of REDCap randomization table CSVs
}
\description{
Export randomization allocation data for a project from the REDCap
  randomization tables but in a form that reflects the allocation tables
  REDCap requests for import
}
\examples{
\dontrun{
allocations <- export_allocation_tables_from_project(
  conn = source_conn,
  project_id_to_export = source_project_id
)
}
}
