% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_email.R
\name{get_bad_emails_from_individual_emails}
\alias{get_bad_emails_from_individual_emails}
\title{Scrape an inbox for bad email addresses in bounce messages}
\usage{
get_bad_emails_from_individual_emails(
  username,
  password,
  url = "imaps://outlook.office365.com",
  messages_since_date
)
}
\arguments{
\item{username}{The username of the IMAP mailbox}

\item{password}{The password of the IMAP mailbox}

\item{url}{The IMAP URL of the host that houses the mailbox}

\item{messages_since_date}{The sent date of the oldest message that should be inspected}
}
\value{
A dataframe of bounced email addresses
\itemize{
  \item{\code{email}}{character email address the bounced}
}
}
\description{
Connect to an imap mailbox, identify bad email addresses referenced in bounce
messages sent after `messages_since_date`, and extract the data from those emails.
}
\examples{
\dontrun{
get_bad_emails_from_individual_emails(
  username = "jdoe",
  password = "jane_does_password",
  url ="imaps://outlook.office365.com",
  messages_since_date = as.Date("2022-01-01", format = "\%Y-\%m-\%d")
  )
}
}
