% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listserv.R
\name{get_bad_emails_from_listserv_digest}
\alias{get_bad_emails_from_listserv_digest}
\title{Enumerate bad email addresses described in LISTSERV email digests}
\usage{
get_bad_emails_from_listserv_digest(
  username,
  password,
  url = "imaps://outlook.office365.com",
  messages_since_date
)
}
\arguments{
\item{username}{The username of the IMAP mailbox}

\item{password}{The password of the IMAP mailbox}

\item{url}{The IMAP URL of the host that houses the mailbox}

\item{messages_since_date}{The sent date of the oldest message that should be inspected}
}
\value{
A dataframe of bad email addresses
\itemize{
  \item email - a column of bad email address
}
}
\description{
Connect to an imap mailbox, identify LISTSERV digest emails sent
after `messages_since_date`, and extract bounced email addresses
from those digest messages.
}
\examples{
\dontrun{
get_bad_emails_from_listserv_digest(
  username = "jdoe",
  password = "jane_does_password",
  url ="imaps://outlook.office365.com",
  messages_since_date = as.Date("2022-01-01", format = "\%Y-\%m-\%d")
  )
}
}
