% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_bad_emails_from_listserv_digest_test_output}
\alias{get_bad_emails_from_listserv_digest_test_output}
\title{get_bad_emails_from_listserv_digest_test_output}
\format{
A data frame with 7 rows and 1 variables:
\describe{
  \item{\code{email}}{character bad email address found by querying listserv error digests}
}
}
\usage{
get_bad_emails_from_listserv_digest_test_output
}
\description{
example output from get_bad_emails_from_listserv_digest that aligns with the related tests
}
\details{
DETAILS
}
\keyword{datasets}
