% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_current_time}
\alias{get_current_time}
\title{Fetches the current time in system time zone}
\usage{
get_current_time()
}
\value{
A duration object representing the current time
}
\description{
Fetches the current time in system time zone
}
\examples{
\dontrun{
get_current_time()
}
}
