% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hipaa_disclosure_log_from_ehr_fhir_logs.R
\name{get_hipaa_disclosure_log_from_ehr_fhir_logs}
\alias{get_hipaa_disclosure_log_from_ehr_fhir_logs}
\title{get_hipaa_disclosure_log_from_ehr_fhir_logs}
\usage{
get_hipaa_disclosure_log_from_ehr_fhir_logs(
  conn,
  ehr_id = NA_real_,
  start_date = as.Date(NA)
)
}
\arguments{
\item{conn}{a DBI connection object to the REDCap database}

\item{ehr_id}{the REDCap EHR_ID for the EHR of interest (optional)}

\item{start_date}{The first date from which we should return results (optional)}
}
\value{
A dataframe suitable for generating a HIPAA disclosure log
}
\description{
Read a data needed for a HIPAA disclosure log from a REDCap database
given a DBI connection object to the REDCap database and some optional
parameters to narrow the returned result.
}
\examples{
\dontrun{
library(tidyverse)
library(lubridate)
library(REDCapR)
library(dotenv)
library(redcapcustodian)
library(DBI)
library(RMariaDB)

init_etl("export_fhir_traffic_log")
conn <- connect_to_redcap_db()

get_hipaa_disclosure_log_from_ehr_fhir_logs(conn)
}
}
