% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutional.R
\name{get_institutional_person_data}
\alias{get_institutional_person_data}
\title{A template function for fetching authoritative email address data
and other institutional data}
\usage{
get_institutional_person_data(user_ids = c(NA_character_))
}
\arguments{
\item{user_ids}{a optional vector of REDCap user IDs to be used in a query
against the institutional data}
}
\value{
A Dataframe
\itemize{
  \item user_id - a column of redcap user_ids / institutional IDs
  \item email - a column of with the authoritative email address for user_id
  \item ... - Additional columns are allowed in the return data frame
}
}
\description{
A template function for fetching authoritative email address data
and other institutional data
}
\examples{
redcap_users <- c("jane_doe", "john_q_public")
get_institutional_person_data(user_ids = redcap_users)
}
