% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_job_duration}
\alias{get_job_duration}
\title{Provide the exact length of the time span between start time and end time}
\usage{
get_job_duration(start_time, end_time)
}
\arguments{
\item{start_time, }{a lubridate::duration object representing the start time}

\item{end_time, }{a lubridate::duration object representing the end time}
}
\value{
the exact length of the time span between start time and end time
}
\description{
Provide the exact length of the time span between start time and end time
}
\examples{
\dontrun{get_job_duration(get_script_run_time(),get_current_time()}

}
