% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_scope_var}
\alias{get_package_scope_var}
\title{Get the value from the redcapcustodian.env environment}
\usage{
get_package_scope_var(key)
}
\arguments{
\item{key}{The identifying string to lookup}
}
\description{
Get the value from the redcapcustodian.env environment
}
\examples{
\dontrun{
  get_package_scope_var("hello")
}
}
