% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_project_instance}
\alias{get_project_instance}
\title{Fetches the package-scoped value of project_instance}
\usage{
get_project_instance()
}
\description{
Fetches the package-scoped value of project_instance
}
