% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redcap_credentials.R
\name{get_redcap_credentials}
\alias{get_redcap_credentials}
\title{Retrieve REDCap Credentials Based on Specified Parameters}
\usage{
get_redcap_credentials(
  project_pid = NA,
  server_short_name = NA,
  project_short_name = NA,
  username = NA
)
}
\arguments{
\item{project_pid}{Optional project ID for filtering.}

\item{server_short_name}{Optional server short name for filtering.}

\item{project_short_name}{Optional project short name for filtering.}

\item{username}{Optional username for filtering.}
}
\value{
A dataframe of filtered REDCap credentials, including a 'url' column added for convenience.
}
\description{
Fetches REDCap credentials from the CREDENTIALS_DB, allowing filtering based on
project ID, server short name, project short name, and username. At least one filtering
criterion must be provided.
}
\examples{
\dontrun{
  source_credentials <- get_redcap_credentials(project_pid = "123")
  prod_credentials <- get_redcap_credentials(server_short_name = "prod")
  target_credentials <- prod_credentials |>
    filter(str_detect(project_name, "biospecimens"))
}

}
