% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_redcap_db_connection}
\alias{get_redcap_db_connection}
\title{Get connection to the primary REDCap database}
\usage{
get_redcap_db_connection()
}
\value{
The existing connection object to the REDCap database
}
\description{
Get connection to the primary REDCap database
}
\examples{
\dontrun{
conn <- get_redcap_db_connection()
}
}
