% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_redcap_email_revisions}
\alias{get_redcap_email_revisions}
\title{Get redcap user email revisions}
\usage{
get_redcap_email_revisions(bad_redcap_user_emails, person)
}
\arguments{
\item{bad_redcap_user_emails}{bad redcap user email data}

\item{person}{institutional person data keyed on user_id}
}
\value{
a dataframe with these columns:
\itemize{
  \item ui_id - row_id of table
  \item username - REDCap username"
  \item email_field_name - the name of the column containing the email address
  \item corrected_email - the corrected email address in email_field_name
}
}
\description{
Get redcap user email revisions
}
\examples{
\dontrun{
conn <- dbConnect(RSQLite::SQLite(), dbname = ":memory:")
bad_emails <- get_bad_redcap_user_emails()
persons <- get_institutional_person_data(conn)
email_revisions <- get_redcap_email_revisions(bad_emails, persons)
}
}
