% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_redcap_email_revisions_test_data}
\alias{get_redcap_email_revisions_test_data}
\title{get_redcap_email_revisions_test_data}
\format{
A list of 3 tibbles:
\describe{
  \item{\code{bad_redcap_user_emails}}{tibble of bad email address and redcap usernames}
  \item{\code{person}}{tibble of corrected email addresses and the corresponding usernames}
  \item{\code{output}}{tibble of output from get_redcap_email_revisions}
}
}
\usage{
get_redcap_email_revisions_test_data
}
\description{
A list of test inputs and outputs for get_redcap_email_revisions
}
\details{
DETAILS
}
\keyword{datasets}
