% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_redcap_emails}
\alias{get_redcap_emails}
\title{Get all user emails in redcap_user_information as tall data}
\usage{
get_redcap_emails(conn)
}
\arguments{
\item{conn}{A DBI Connection object}
}
\value{
a list of 2 dataframes:
\itemize{
  \item wide, relevant email columns from redcap_user_information
  \item tall, wide data pivoted to include email_field_name and email columns
}
}
\description{
Get all user emails in redcap_user_information as tall data
}
\examples{
\dontrun{
conn <- dbConnect(RSQLite::SQLite(), dbname = ":memory:")
create_test_table(conn, "redcap_user_information")
get_redcap_emails(conn)
}
}
