% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_redcap_emails_output}
\alias{get_redcap_emails_output}
\title{get_redcap_emails_output}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{\code{ui_id}}{double primary key}
  \item{\code{username}}{character redcap username}
  \item{\code{email_field_name}}{character column name in redcap_user_information}
  \item{\code{email}}{character email value in the column named in `email_field_name`}
}
}
\usage{
get_redcap_emails_output
}
\description{
user email addresses from the table redcap_user_information.
Addresses have been pivoted taller into a single `email` column.
}
\details{
DETAILS
}
\keyword{datasets}
