% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_redcap_emails_test_data}
\alias{get_redcap_emails_test_data}
\title{get_redcap_emails_test_data}
\format{
An object of class \code{list} of length 2.
}
\usage{
get_redcap_emails_test_data
}
\value{
a list of 2 dataframes:
\itemize{
  \item wide, relevant email columns from redcap_user_information
  \item tall, wide data pivoted to include email_field_name and email columns
}
}
\description{
test data to mock get_redcap_emails
}
\details{
DETAILS
}
\keyword{datasets}
