% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_script_run_time}
\alias{get_script_run_time}
\title{Fetches the package-scoped value of script_run_time}
\usage{
get_script_run_time()
}
\description{
Fetches the package-scoped value of script_run_time
}
\examples{
get_script_run_time()
}
