% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_database_comparison.R
\name{get_table_checksum}
\alias{get_table_checksum}
\title{get_table_checksum}
\usage{
get_table_checksum(table_name, conn)
}
\arguments{
\item{table_name}{The name of a table in the database described by `conn`}

\item{conn}{A DBI Connection object to a MySQL database}
}
\value{
A one row dataframe with these columns:
\describe{
  \item{host}{MySQL host name found on the DBI connection object}
  \item{database_name}{MySQL database name returned by the query}
  \item{table}{MySQL table name}
  \item{checksum}{checksum returned by "CHECKSUM TABLE <table_name>"}
  \item{elapsed_time}{elapsed time required for the query and fetch of the checksum}
}
}
\description{
Fetch the checksum of a MySQL table accessed via DBI connection object.
}
\examples{
\dontrun{
get_table_checksum("redcap_user_information", rc_conn)
}
}
