% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{get_test_table_names}
\alias{get_test_table_names}
\title{Provides a list of table names which have schema and data files as part of the package}
\usage{
get_test_table_names()
}
\value{
A list of table names which have schema and data files as part of the package
}
\description{
Provides a list of table names which have schema and data files as part of the package
}
\examples{
get_test_table_names()
}
