% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{init_log_con}
\alias{init_log_con}
\title{Initialize the connection to the log db and set redcapcustodian.env$log_con}
\usage{
init_log_con(drv = RMariaDB::MariaDB())
}
\arguments{
\item{drv, }{an object that inherits from DBIDriver (e.g. RMariaDB::MariaDB()), or an existing DBIConnection object (in order to clone an existing connection).}
}
\description{
Initialize the connection to the log db and set redcapcustodian.env$log_con
}
\examples{
\dontrun{
 # use a sqlite db instead
 init_log_con(drv = RSQLite::SQLite())
}
}
