% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_db_con}
\alias{is_db_con}
\title{Check if the provided connection is a DBI connection object}
\usage{
is_db_con(con)
}
\arguments{
\item{con}{a DBI connection}
}
\value{
The result of the test
}
\description{
Check if the provided connection is a DBI connection object
}
\examples{
\dontrun{
 conn = connect_to_local_db()
 is_db_con(
   con = conn
 )
}
}
