% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_on_ci}
\alias{is_on_ci}
\title{Check if "CI" environment variable is set to TRUE}
\usage{
is_on_ci()
}
\value{
TRUE/FALSE
}
\description{
Check if "CI" environment variable is set to TRUE
}
\examples{
\dontrun{
  is_on_ci()
}

}
