% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_job_debug}
\alias{log_job_debug}
\title{Log a job debug entry}
\usage{
log_job_debug(summary)
}
\arguments{
\item{summary, }{the job summary}
}
\description{
Log a job debug entry
}
\examples{
\dontrun{
 log_job_debug(
   summary = "Job debug step"
 )
}
}
