% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_job_failure}
\alias{log_job_failure}
\title{Log a failed job run}
\usage{
log_job_failure(summary)
}
\arguments{
\item{summary, }{the job summary}
}
\description{
Log a failed job run
}
\examples{
\dontrun{
 log_job_failure(
   summary = "Job failed"
 )
}
}
