% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_job_success}
\alias{log_job_success}
\title{Log a successful job run}
\usage{
log_job_success(summary)
}
\arguments{
\item{summary, }{the job summary}
}
\description{
Log a successful job run
}
\examples{
\dontrun{
 log_job_success(
   summary = "Job succeeded"
 )
}
}
