% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_data_to_redcap_project.R
\name{mirror_data_to_redcap_project}
\alias{mirror_data_to_redcap_project}
\title{mirror_data_to_redcap_project}
\usage{
mirror_data_to_redcap_project(df_to_mirror, table_name, path_to_env_file, ...)
}
\arguments{
\item{df_to_mirror}{the dataframe to be imported into the REDCap project}

\item{table_name}{the table name}

\item{path_to_env_file}{the env file path}

\item{...}{Additional parameters that will be passed to \code{\link[REDCapR]{redcap_write}}}
}
\value{
Same values returned as \code{\link[REDCapR]{redcap_write}}
}
\description{
mirror a dataframe to a REDCap project
}
\examples{
\dontrun{
mirror_data <- mirror_data_to_redcap_project(df, "table_name", "./folder/path/.env")
}

}
