% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_creation_descriptions}
\alias{project_creation_descriptions}
\title{project_creation_descriptions}
\format{
A character vector with 7 elements
}
\usage{
project_creation_descriptions
}
\description{
A character vector of the descriptions used in the redcap_log_event table
  to describe the different ways a REDCap Project can be created
}
\details{
DETAILS
}
\keyword{datasets}
