% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_life_cycle_descriptions}
\alias{project_life_cycle_descriptions}
\title{project_life_cycle_descriptions}
\format{
A charecter vector with 24 elements
}
\usage{
project_life_cycle_descriptions
}
\description{
A character vector of the descriptions used in the redcap_log_event table
  to describe the different stages in the life cycle of a REDCap Project
}
\details{
DETAILS
}
\keyword{datasets}
