% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_purpose_labels}
\alias{project_purpose_labels}
\title{project_purpose_labels}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{\code{id}}{double primary key}
  \item{\code{project_purpose}}{character redcap project purpose}
}
}
\usage{
project_purpose_labels
}
\description{
A tibble project purpose IDs and project purposes that reflect their
  meaning as used in the `purpose` column of the `redcap_projects` table
}
\details{
DETAILS
}
\keyword{datasets}
