% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_purpose_other_research_labels}
\alias{project_purpose_other_research_labels}
\title{project_purpose_other_research_labels}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{\code{id}}{double primary key}
  \item{\code{project_purpose_other_research}}{character redcap purpose other research}
}
}
\usage{
project_purpose_other_research_labels
}
\description{
A tibble project purpose other IDs and labels that reflect their
  meaning as used in the `purpose_other` column of the `redcap_projects` table
}
\details{
DETAILS
}
\keyword{datasets}
