% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_status_labels}
\alias{project_status_labels}
\title{project_status_labels}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{\code{id}}{double primary key}
  \item{\code{project_status}}{character redcap project status}
}
}
\usage{
project_status_labels
}
\description{
A tibble of project status IDs and project statuses that reflect their
  meaning as used in the `status` column of the `redcap_projects` table
}
\details{
DETAILS
}
\keyword{datasets}
