% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quit_non_interactive_run}
\alias{quit_non_interactive_run}
\title{Quit a non interactive R session}
\usage{
quit_non_interactive_run()
}
\description{
Quit a non interactive R session
}
\examples{
\dontrun{
  quit_non_interactive_run()
}
}
