% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project_data.R
\name{read_project_data}
\alias{read_project_data}
\title{read_project_data Read data from a REDCap project}
\usage{
read_project_data(..., conn, project_pid, server_short_name = as.character(NA))
}
\arguments{
\item{...}{parameters to pass on to `REDCapR::redcap_read`}

\item{conn}{Connection object to the credentials database}

\item{project_pid}{the project PID in the REDCap system.
This string will be used to search through a REDCap custodian
credentials database to locate the `token` and `redcap_uri`}

\item{server_short_name}{an optional name of the server that
houses the REDCap project of interest. This will prevent
project PID clashes.}
}
\value{
a dataframe of data read from a REDCap project
}
\description{
read_project_data Read data from a REDCap project
}
\examples{
\dontrun{
library(redcapcustodian)
library(DBI)

dotenv::load_dot_env("testing.env")
init_etl("dummy")

support_data <- read_project_data(
  conn = DBI::dbConnect(RSQLite::SQLite(), Sys.getenv("CREDENTIALS_DB")),
  project_pid = Sys.getenv("MY_PROJECT_PID"))
}
}
