% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{send_email}
\alias{send_email}
\title{A wrapper function that sends an email (via sendmailR) reporting the outcome of another function}
\usage{
send_email(
  email_body,
  email_subject = "",
  email_to = "",
  email_cc = "",
  email_from = ""
)
}
\arguments{
\item{email_body}{The contents of the email}

\item{email_subject}{The subject line of the email}

\item{email_to}{The email addresses of the primary recipient(s), separate recipient addresses with spaces}

\item{email_cc}{The email addresses of cc'd recipient(s), separate recipient addresses with spaces}

\item{email_from}{The email addresses of the sender}
}
\value{
No returned value
}
\description{
A wrapper function that sends an email (via sendmailR) reporting the outcome of another function
}
\examples{

\dontrun{
message <- paste("Failed REDCap data import to", project_title,
                  "\nThe reason given was:", error_message)

email_subject <- paste("FAILED |", script_name, "|",
                        Sys.getenv("INSTANCE"), "|", script_run_time)

send_email(message, email_subject)
}
}
