% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_package_scope_var}
\alias{set_package_scope_var}
\title{Assign a value to the redcapcustodian.env environment, retrievable with \code{\link{get_package_scope_var}}}
\usage{
set_package_scope_var(key, value)

set_package_scope_var(key, value)
}
\arguments{
\item{key}{The identifying string to store as the lookup key}

\item{value}{The value to store}
}
\description{
Assign a value to the redcapcustodian.env environment, retrievable with \code{\link{get_package_scope_var}}

Assign a value to the redcapcustodian.env environment, retrievable with \code{\link{get_package_scope_var}}
}
\examples{
\dontrun{
  set_package_scope_var("hello", "world")
  hello <- get_package_scope_var("hello")
}
\dontrun{
  set_package_scope_var("hello", "world")
  hello <- get_package_scope_var("hello")
}
}
