% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_management.R
\name{set_project_api_token}
\alias{set_project_api_token}
\title{Generate an API token for a REDCap project and assign it to a specified user}
\usage{
set_project_api_token(conn, username, project_id)
}
\arguments{
\item{conn}{a DBI database connection, such as that from \code{\link{get_redcap_db_connection}}}

\item{username}{a REDCap username \cr
This user must already have access to the project, i.e. they must appear in the project's User Rights}

\item{project_id}{The project id of the project for which you wish to generate a token}
}
\value{
nothing
}
\description{
Generate an API token for a REDCap project and assign it to a specified user
}
\examples{
\dontrun{
  conn <- get_redcap_db_connection()
  my_new_token <- set_project_api_token(conn, "admin", 15)
}
}
