% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_project_name}
\alias{set_project_name}
\title{Sets the package-scoped value of project_name}
\usage{
set_project_name(project_name = "")
}
\arguments{
\item{project_name}{Defaults to NULL. If provided and not NULL, this value is used.
If NULL, the function attempts to fetch the value from the environment variable.}
}
\value{
the package-scoped value of project_name
}
\description{
Sets the package-scoped value of project_name
}
\examples{
\dontrun{
project_name <- set_project_name()
project_name <- set_project_name("project_name")
}

}
