% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_script_name}
\alias{set_script_name}
\title{Assigns package-scoped script_name. \cr
By default this is sourced from the focused RStudio window or
the calling command (e.g. Rscript script_name.R)}
\usage{
set_script_name(script_name = "")
}
\arguments{
\item{script_name}{optional arg to override the calling script}
}
\description{
Assigns package-scoped script_name. \cr
By default this is sourced from the focused RStudio window or
the calling command (e.g. Rscript script_name.R)
}
