% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_script_run_time}
\alias{set_script_run_time}
\title{Sets the package-scoped value of script_run_time}
\usage{
set_script_run_time(fake_runtime = lubridate::NA_POSIXct_)
}
\arguments{
\item{fake_runtime}{An asserted script run time}
}
\value{
the package-scoped value of script_run_time
}
\description{
Sets the package-scoped value of script_run_time
}
\examples{
set_script_run_time()
set_script_run_time(fake_runtime =
                    as.POSIXct("2021-02-23 02:23:00",
                               tz="",
                               format="\%Y-\%m-\%d \%H:\%M:\%OS")
                   )
}
