% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_management.R
\name{set_super_api_token}
\alias{set_super_api_token}
\title{Generate and set a Super API token for a provided REDCap user}
\usage{
set_super_api_token(conn, username)
}
\arguments{
\item{conn}{a DBI database connection, such as that from \code{\link{get_redcap_db_connection}}}

\item{username}{a REDCap username}
}
\value{
The newly created super token
}
\description{
Generate and set a Super API token for a provided REDCap user
}
\examples{
\dontrun{
  conn <- get_redcap_db_connection()
  my_new_super_token <- set_super_api_token(conn, "admin")
}
}
