% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{suspend_users_with_no_primary_email}
\alias{suspend_users_with_no_primary_email}
\title{Suspends users with no primary email in redcap_user_information}
\usage{
suspend_users_with_no_primary_email(conn)
}
\arguments{
\item{conn}{A DBI Connection object}
}
\description{
Suspends users with no primary email in redcap_user_information
}
\examples{
\dontrun{
suspend_users_with_no_primary_email(conn)
}
}
