% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_instance.R
\name{sync_metadata}
\alias{sync_metadata}
\title{Sync data dictionary of a source project to a target project using credential objects}
\usage{
sync_metadata(
  source_credentials,
  target_credentials,
  strip_action_tags = FALSE
)
}
\arguments{
\item{source_credentials}{A dataframe containing the following columns:
\itemize{
  \item redcap_uri - The uri for the API endpoint of a REDCap host
  \item token - The REDCap API token for a specific project
}
This dataframe should contain credentials for the project you wish to copy from}

\item{target_credentials}{A dataframe containing the following columns:
\itemize{
  \item redcap_uri - The uri for the API endpoint of a REDCap host
  \item token - The REDCap API token for a specific project
}
This dataframe should contain credentials for the project you wish to overwrite}

\item{strip_action_tags}{Optional toggle to remove action tags, useful for porting to a development environment; defaults to FALSE}
}
\value{
nothing
}
\description{
Sync data dictionary of a source project to a target project using credential objects
}
\examples{
\dontrun{
  source_credentials <- REDCapR::retrieve_credential_local(
    path_credential = "source_credentials.csv",
    project_id = 31
  )

  target_credentials <- REDCapR::retrieve_credential_local(
    path_credential = "target_credentials.csv",
    project_id = 25
  )

  sync_metadata(source_credentials, target_credentials)
}
}
\seealso{
\code{\link[REDCapR]{retrieve_credential}}, \code{\link{scrape_user_api_tokens}}
 \code{vignette("credential-scraping", package = "redcapcustodian")}
}
