% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{sync_table}
\alias{sync_table}
\title{Write to a MySQL Database using the result of \code{\link{dataset_diff}}}
\usage{
sync_table(
  conn,
  table_name,
  primary_key,
  data_diff_output,
  insert = F,
  update = T,
  delete = F
)
}
\arguments{
\item{conn}{a DBI database connection}

\item{table_name}{name of the table to write to}

\item{primary_key}{name of the primary key (or vector of multiple keys) of the table to write to}

\item{data_diff_output}{list of dataframes returned by \code{\link{dataset_diff}}}

\item{insert}{boolean toggle to use the insert dataframe to insert rows in \code{table_name}}

\item{update}{boolean toggle to use the updates dataframe to update rows in \code{table_name}}

\item{delete}{boolean toggle to use the delete dataframe to delete rows in \code{table_name}}
}
\value{
a named list with these values:
\itemize{
  \item insert_n - the number of rows inserted to \code{table_name}
  \item update_n - the number of rows updated in \code{table_name}
  \item delete_n - the number of rows deleted in \code{table_name}
}
}
\description{
Write to a MySQL Database using the result of \code{\link{dataset_diff}}
}
\examples{
\dontrun{
conn <- connect_to_redcap_db()

 ...

diff_output <- dataset_diff(
  source = updates,
  source_pk = "id",
  target = original_table,
  target_pk = "id"
)

sync_table(
  conn = con,
  table_name = table_name,
  primary_key = primary_key,
  data_diff_output = diff_output
)
}
}
