% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{sync_table_2}
\alias{sync_table_2}
\title{Write to a MySQL Database based on the diff of source and target datasets.}
\usage{
sync_table_2(
  conn,
  table_name,
  source,
  source_pk,
  target,
  target_pk,
  insert = F,
  update = T,
  delete = F
)
}
\arguments{
\item{conn}{a DBI database connection}

\item{table_name}{name of the table to write to}

\item{source}{- a dataframe with content that needs to be reflected in target}

\item{source_pk}{- the primary key of source}

\item{target}{- a data frame that needs to reflect source}

\item{target_pk}{- the primary key of target}

\item{insert}{boolean toggle to use the insert dataframe to insert rows in \code{table_name}}

\item{update}{boolean toggle to use the updates dataframe to update rows in \code{table_name}}

\item{delete}{boolean toggle to use the delete dataframe to delete rows in \code{table_name}}
}
\value{
a named list with these values:
\itemize{
  \item insert_records - a dataframe of inserts
  \item update_records - a dataframe of updates
  \item delete_records - a dataframe of deletions
  \item insert_n - the number of rows inserted to \code{table_name}
  \item update_n - the number of rows updated in \code{table_name}
  \item delete_n - the number of rows deleted in \code{table_name}
}
}
\description{
Write to a MySQL Database based on the diff of source and target datasets.
}
\examples{
\dontrun{
conn <- connect_to_redcap_db()

 ...

sync_table_2(
  conn = con,
  table_name = table_name,
  source = updates,
  source_pk = "id",
  target = original_table,
  target_pk = "id"
)
}
}
