% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sync_table_test_user_data_result}
\alias{sync_table_test_user_data_result}
\title{sync_table_test_user_data_result}
\format{
A data frame with 4 rows and 3 variables:
\describe{
  \item{\code{ui_id}}{double primary key}
  \item{\code{username}}{character redcap username}
  \item{\code{user_email}}{character primary email address}
}
}
\usage{
sync_table_test_user_data_result
}
\description{
output for sync_table when dataset_diff_test_user_data is used as the input
}
\details{
DETAILS
}
\keyword{datasets}
