% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{update_redcap_email_addresses}
\alias{update_redcap_email_addresses}
\title{Updates bad redcap email addresses in redcap_user_information}
\usage{
update_redcap_email_addresses(conn, redcap_email_revisions)
}
\arguments{
\item{redcap_email_revisions}{a df returned by  \code{\link{get_redcap_email_revisions}}

\dontrun{
conn <- dbConnect(RSQLite::SQLite(), dbname = ":memory:")
update_redcap_email_addresses(conn, redcap_email_revisions)
}}
}
\description{
Updates bad redcap email addresses in redcap_user_information
}
