% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{update_redcap_email_addresses_test_data}
\alias{update_redcap_email_addresses_test_data}
\title{update_redcap_email_addresses_test_data}
\format{
A list of 1 variables:
\describe{
  \item{\code{output}}{tibble of email data read backfrom redcap_user_information}
}
}
\usage{
update_redcap_email_addresses_test_data
}
\description{
A list of test inputs and outputs for update_redcap_email_addresses
}
\details{
DETAILS
}
\keyword{datasets}
