% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{user_rights_test_data}
\alias{user_rights_test_data}
\title{user_rights_test_data}
\format{
A named list of 3 dataframes:
\describe{
  \item{\code{redcap_user_information}}{REDCap Core table}
  \item{\code{redcap_user_rights}}{REDCap Core table}
  \item{\code{redcap_user_roles}}{REDCap Core table}
}
}
\usage{
user_rights_test_data
}
\description{
A named list of dataframes used to test the functions written to manage user rights
}
\details{
DETAILS
}
\keyword{datasets}
