% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{verify_log_connectivity}
\alias{verify_log_connectivity}
\title{Attempts to connect to the DB using all LOG_DB_* environment variables. Returns an empty list if a connection is established, returns an `error_list` entry otherwise.}
\usage{
verify_log_connectivity(drv)
}
\arguments{
\item{drv, }{an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).}
}
\value{
An `error_list` entry
}
\description{
Attempts to connect to the DB using all LOG_DB_* environment variables. Returns an empty list if a connection is established, returns an `error_list` entry otherwise.
}
\examples{
\dontrun{
 verify_log_connectivity()
}
}
