% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{verify_log_dependencies}
\alias{verify_log_dependencies}
\title{Verifies all dependencies required to write log entries.}
\usage{
verify_log_dependencies(drv = RMariaDB::MariaDB())
}
\arguments{
\item{drv, }{an object that inherits from DBIDriver (e.g. RMariaDB::MariaDB()), or an existing DBIConnection object (in order to clone an existing connection).}
}
\value{
A list of `error_list` entries.
}
\description{
Verifies all dependencies required to write log entries.
}
\examples{
\dontrun{
 verify_log_dependencies(
     drv = RMariaDB::MariaDB()
 )
}
}
