% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{verify_log_env_variables}
\alias{verify_log_env_variables}
\title{Verifies all required environment variables are set. Returns an empty list if all necessary environment variables are set, returns a list of errors otherwise.}
\usage{
verify_log_env_variables()
}
\value{
A list of `error_list` entries.
}
\description{
Verifies all required environment variables are set. Returns an empty list if all necessary environment variables are set, returns a list of errors otherwise.
}
\examples{
\dontrun{
 verify_log_env_variables()
}
}
