% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{write_allocations}
\alias{write_allocations}
\title{write_allocations}
\usage{
write_allocations(project_status_to_write, allocations, target_directory = ".")
}
\arguments{
\item{project_status_to_write}{- the value of project_status to export.
Use 0 for development. Use 1 for Production}

\item{allocations}{- the dataframe of randomization allocation data as
exported by `export_allocation_tables_from_project`}

\item{target_directory}{- the directory into which the function should write the files}
}
\value{
the full path to the allocations file
}
\description{
Write the development or production randomization allocation table in
  the same form in which it was loaded.
}
\examples{
\dontrun{
# get and print importable allocations if we need them for reference
allocations <- export_allocation_tables_from_project(
  conn = source_conn,
  project_id_to_export = source_project_id
)

# write both files
walk(c(0,1), write_allocations, allocations, "output")
}
}
