% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{write_error_log_entry}
\alias{write_error_log_entry}
\title{Write an error log entry}
\usage{
write_error_log_entry(con, target_db_name, table_written = NULL, df, pk_col)
}
\arguments{
\item{con, }{a DB connection}

\item{target_db_name, }{the database to write to}

\item{table_written, }{the table that was written to}

\item{df, }{the data to write}

\item{pk_col, }{the dataframe col to use as the primary_key}
}
\description{
Write an error log entry
}
\examples{
\dontrun{
 write_error_log_entry(
   conn = con,
   target_db_name = rc_case,
   table_written = "cases"
   df = data_written,
   pk_col = "record_id",
 )
}
}
