% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{write_success_job_log_entry}
\alias{write_success_job_log_entry}
\title{Write an success job log entry}
\usage{
write_success_job_log_entry(con, job_duration, job_summary)
}
\arguments{
\item{con, }{a DB connection}

\item{job_duration, }{the duration of the job in seconds}

\item{job_summary, }{the summary of the job performed as JSON}
}
\description{
Write an success job log entry
}
\examples{
\dontrun{
 write_success_job_log_entry(
   conn = con,
   job_duration = 30,
   job_summary = as.json("Update vaccination status")
 )
}
}
