% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{write_summary_metrics}
\alias{write_summary_metrics}
\title{Format and write summary metrics to the redcap_summary_metrics table in your LOG_DB}
\usage{
write_summary_metrics(
  reporting_period_start,
  reporting_period_end,
  metric_type,
  metric_dataframe,
  conn = NULL
)
}
\arguments{
\item{reporting_period_start}{a datetime object, e.g. ymd_hms("2022-11-01 00:00:00")}

\item{reporting_period_end}{a datetime object, e.g. ymd_hms("2022-12-01 00:00:00")}

\item{metric_type}{a character string representing the metric type, e.g. "flux", "state"}

\item{metric_dataframe}{A wide data frame of key-value pairs with a single row of data}

\item{conn}{A DBI connection object to the database that holds the
`redcap_summary_metrics` table. Can be left as NULL if the connection is available
on the package scope var "log_con".}
}
\value{
nothing
}
\description{
Format and write summary metrics to the redcap_summary_metrics table in your LOG_DB
}
\examples{
\dontrun{
 write_summary_metrics(
   reporting_period_start = ymd_hms("2022-01-01 00:00:00", tz=Sys.getenv("TIME_ZONE")),
   reporting_period_end = ceiling_date(reporting_period_start, "month", change_on_boundary = T),
   metric_type = "state",
   metric_dataframe = my_cool_df,
   conn = my_conn
 )
}
}
