% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{connect_to_redcap_db}
\alias{connect_to_redcap_db}
\title{Connect to the Primary REDCap MySQL Database
Assigns package-scoped conn}
\usage{
connect_to_redcap_db()
}
\value{
An S4 object. Run ?dbConnect for more information
}
\description{
Connect to the Primary REDCap MySQL Database
Assigns package-scoped conn
}
\examples{
\dontrun{
conn <- connect_to_redcap_db()
}
}
