% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_to_redcap_dictionary.R
\name{dataframe_to_redcap_dictionary}
\alias{dataframe_to_redcap_dictionary}
\title{dataframe_to_redcap_dictionary}
\usage{
dataframe_to_redcap_dictionary(
  df,
  form_name,
  write_to_csv = FALSE,
  filename = NULL
)
}
\arguments{
\item{df}{the dataframe to generate the data dictionary for}

\item{form_name}{the form name to display in REDCap}

\item{write_to_csv}{If TRUE will write the datadictionary to a csv.}

\item{filename}{A string specifying the filename for the CSV.}
}
\value{
A redcap data dictionary
}
\description{
Create a REDCap data dictionary for a dataframe
}
\examples{
\dontrun{

df <- data.frame(
  numeric_col = c(1, 2, 3),
  integer_col = as.integer(5.9),
  character_col = c("a", "b", "c"),
  date_col = as.Date("2011-03-07"),
  date_time_col = as.POSIXct(strptime("2011-03-27 01:30:00", "\%Y-\%m-\%d \%H:\%M:\%S"))
)

redcap_data_dictionary <- dataframe_to_redcap_dictionary(df, "test")
redcap_data_dictionary <- dataframe_to_redcap_dictionary(df, "test", TRUE, "<output_path>.csv")
}
}
