% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_redcap_emails}
\alias{get_redcap_emails}
\title{Get all user emails in redcap_user_information as tall data}
\usage{
get_redcap_emails(conn)
}
\arguments{
\item{conn}{A DBI Connection object}
}
\value{
a dataframe with these columns:
\itemize{
  \item ui_id - row_id of table
  \item username - REDCap username"
  \item email_field_name - the name of the column containing the email address
  \item email - the email address in email_field_name
}
}
\description{
Get all user emails in redcap_user_information as tall data
}
\examples{
\dontrun{
conn <- dbConnect(RSQLite::SQLite(), dbname = ":memory:")
create_test_table(conn, "redcap_user_information")
get_redcap_emails(conn)
}
}
