% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{log_event_tables}
\alias{log_event_tables}
\title{log_event_tables}
\format{
A vector with 9 elements
}
\usage{
log_event_tables
}
\description{
A vector of the names of the 9 redcap log event tables
}
\details{
DETAILS
}
\keyword{datasets}
