% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_data_to_redcap_project.R
\name{mirror_data_to_redcap_project}
\alias{mirror_data_to_redcap_project}
\title{Mirror a dataframe to a REDCap project}
\usage{
mirror_data_to_redcap_project(
  df_to_mirror,
  record_id_col,
  project_name,
  path_to_env_file,
  ...
)
}
\arguments{
\item{df_to_mirror}{Dataframe to be imported into the REDCap project.}

\item{record_id_col}{Column in the dataframe that uniquely identifies each record.}

\item{project_name}{REDCap project name.}

\item{path_to_env_file}{Path to the ".env" credentials file.}

\item{...}{Additional parameters passed to \code{\link[REDCapR]{redcap_write}}.}
}
\value{
Same values as returned by \code{\link[REDCapR]{redcap_write}}.
}
\description{
Mirror a dataframe to a REDCap project
}
\examples{
\dontrun{
df <- data.frame(
  pk_col = c("a1", "a2", "a3"),
  integer_col = c(1, 2, 3),
  numeric_col = 5.9,
  character_col = c("a", "b", "c"),
  date_col = as.Date("2011-03-07"),
  date_time_col = as.POSIXct(strptime("2011-03-27 01:30:00", "\%Y-\%m-\%d \%H:\%M:\%S")),
  email_col = c("test@example.com", "test.2@example.edu", "3test@example.net")
)
mirrored_data <- mirror_data_to_redcap_project(df, "pk_col", "test_project", "<env_path>.env")
}

}
\seealso{
\code{\link{dataframe_to_redcap_dictionary}} to understand the REDCap data dictionary requirements.
The dictionary must be created and uploaded to REDCap for \code{mirror_data_to_redcap_project} to function properly.
}
